refactor laptop resource with latest improvements, add controller promise rejection handling

+++ src/resources/productCategory/accessory/accessory.controller.ts
@@ -425,18 +425,40 @@ const deleteAllAccessoriesHandler = expressAsyncHandler(
 		const uploadedFilesIds = await returnAllAccessoriesUploadedFileIdsService();
 
 		// delete all file uploads associated with all accessories
-		await Promise.all(
+		const deletedFileUploads = await Promise.all(
 			uploadedFilesIds.map(
 				async (fileUploadId) => await deleteFileUploadByIdService(fileUploadId),
 			),
 		);
 
+		if (
+			deletedFileUploads.some(
+				(deletedFileUpload) => deletedFileUpload.deletedCount === 0,
+			)
+		) {
+			response.status(400).json({
+				message: "Some File uploads could not be deleted. Please try again.",
+				resourceData: [],
+			});
+			return;
+		}
+
 		// delete all reviews associated with all accessories
-		await Promise.all(
+		const deletedReviews = await Promise.all(
 			uploadedFilesIds.map(
 				async (fileUploadId) => await deleteAProductReviewService(fileUploadId),
 			),
 		);
+
+		if (
+			deletedReviews.some((deletedReview) => deletedReview.deletedCount === 0)
+		) {
+			response.status(400).json({
+				message: "Some reviews could not be deleted. Please try again.",
+				resourceData: [],
+			});
+			return;
+		}
 
 		// delete all accessories
 		const deleteAccessoriesResult: DeleteResult =
@@ -480,18 +502,40 @@ const deleteAAccessoryHandler = expressAsyncHandler(
 		const uploadedFilesIds = [...accessoryExists.uploadedFilesIds];
 
 		// delete all file uploads associated with all accessories
-		await Promise.all(
+		const deletedFileUploads = await Promise.all(
 			uploadedFilesIds.map(
 				async (fileUploadId) => await deleteFileUploadByIdService(fileUploadId),
 			),
 		);
 
+		if (
+			deletedFileUploads.some(
+				(deletedFileUpload) => deletedFileUpload.deletedCount === 0,
+			)
+		) {
+			response.status(400).json({
+				message: "Some File uploads could not be deleted. Please try again.",
+				resourceData: [],
+			});
+			return;
+		}
+
 		// delete all reviews associated with all accessories
-		await Promise.all(
+		const deletedReviews = await Promise.all(
 			uploadedFilesIds.map(
 				async (fileUploadId) => await deleteAProductReviewService(fileUploadId),
 			),
 		);
+
+		if (
+			deletedReviews.some((deletedReview) => deletedReview.deletedCount === 0)
+		) {
+			response.status(400).json({
+				message: "Some reviews could not be deleted. Please try again.",
+				resourceData: [],
+			});
+			return;
+		}
 
 		// delete accessory by id
 		const deleteAccessoryResult: DeleteResult =
